Moved BaseServerHandler static classes to individual files to comply with CodeFactor.
